perf(frontend): lazy-load Remotion feature player
